check Shader and Texture code, remove from TODO - added extra comment to shader.hpp - changed _textureHandle from uint32_t to uint16_t in texture.hpp

+++ src/backend/texture.hpp
@@ -37,7 +37,7 @@ namespace pen {
         bool persistent = false;
 
         // Texture Data
-        uint32_t _textureHandle;
+        uint16_t _textureHandle;
         uint16_t width, height;
         Vec2 spriteDeform;
         uint8_t textureType;
